refactor: resolved sonarqube issues
